First round of cleanups for ASDimension.mm/h
