+++ AsyncDisplayKit/Layout/ASAbsoluteLayoutElement.h
@@ -9,7 +9,7 @@
 //
 
 #import <AsyncDisplayKit/ASBaseDefines.h>
-#import <AsyncDisplayKit/ASDimension.h>
+#import <AsyncDisplayKit/ASDimensionDeprecated.h>
 
 NS_ASSUME_NONNULL_BEGIN
